+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added the 2026 Sergio Sisternes copyright notice to the root Apache-2.0
-  license.
+- Added the 2026 Sergio Sisternes copyright notice to the README license
+  section.
 
 ## [0.3.10] - 2026-09-10